allow @ after [unplate.begin(NAME)]
The reason for this is that the code

  [unplate.begin(NAME)] """
  my template
  """ [unplate.end]

is parsed with [unplate.begin(NAME)] as an expression followed
by """\nmytemplate\n""" [unplate.end] as another expression
(namely, indexing into a string literal). This places two
expressions on the same logical line, which is illegal.

By placing an @ in between the opening and the string literal:

  [unplate.begin(NAME)] @ """
  my template
  """ [unplate.end]

then the code becomes a valid expression, since @ is a binary
operator.