fix leading/trailing newline stripping; unindent string literals
1) fix leading/trailing newline stripping:

In code like

  if True:
    [unplate.template(NAME)] @ '''
    my template
    ''' [unplate.end]

Unplate would complain because it wants the template
to begin and end with a newline, but this ends with an
indent. Instead, the requirement has been changed to
beginning and ending with blank lines, where a blank line
must contain only whitespace.

2) unindent string literals

In

  if True:
    [unplate.template(NAME)] @ '''
    a
      b
    c
    ''' [unplate.end]

the indentation of the string literal is now stripped.
So the resultant template is now

  "a\n  b\nc\n"

instead of

  "  a\n    b\n  c\n"