Generate whitespace better
This is reverting commit f5e3d1d8ffacdde4972f8b612f52b28b58a8e09b ("preemptively fix a bug")
The indent stack stores accumulated indentation, not indentation deltas.
By this I mean that in the code

  if A:
    if B:
      if C:
        pass

on the pass statement, the indent stack looks like

  ['  ', '    ', '      ']

i.e. 2 spaces, then 4, then 6. This reflects the accumulated
indentation at each indent.
The indent stack does NOT look like

  ['  ', '  ', '  ']

i.e. 2 spaces for each indent. This would reflect how much
each indentation is indeted by.

As such, the code used to generate a new indent up one level
from the current was inappropriate. It was summing the indents,
as if the stack were of the second form, when in fact it just
needed to read off the top of the stack.